chore: create contribution standards
Add a description of contribution standards to the docs. This includes standards for commits and pull requests as well as formatting and linting information. Set up Husky to enforce formatting/linting as well as Conventional Commits commit messages. Add a commitlint job to continuous integration to verify commit messages. Formatting and linting is already checked by CI. This sets a standard to ensure high-quality contributions.