feat: look up nearest way on map click rather than invisible info layer (#81)
The user may click anywhere on the map. The coordinates are sent to the backend to find the nearest way to the click coordinates and the sidebar is opened for that way. This replaces the previous system of loading an invisible 'info layer' with many thousand ways on map load, which had high memory requirements and caused poor browser performance.