feat(routing): tweak routing weight multipliers and some routing style improvements (#95)
Adjust routing weight multipliers to strongly discourage routes where bikes are not allowed, strongly discourage mountain bike trails, and discourage stairs (to a lesser extent than the others). Rename the label on the best route to 'Best Route' from somewhat misleading 'Shortest Route' (since it is not necessarily the shortest by distance, only by cumulative edge weight). Show the blue best route line on top of the alternative green routes, which is important when there is overlap. Move time/distance displays from top right to bottom left of the map to fix overlap with the 'Hide' button for the sidebar.